# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Documentation for Challenge-Go API",
        "title": "Swagger Challenge API",
        "contact": {},
        "version": "2.0"
    },
    "basePath": "/",
    "paths": {
        "/accounts": {
            "get": {
                "description": "List all accounts",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get accounts",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entities.Account"
                            }
                        }
                    },
                    "400": {
                        "description": "Failed to decode"
                    },
                    "404": {
                        "description": "Accounts not found"
                    },
                    "500": {
                        "description": "Unexpected internal server error"
                    }
                }
            },
            "post": {
                "description": "Create an account with the basic information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create an account",
                "parameters": [
                    {
                        "description": "Body",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entities.AccountInput"
                        }
                    }
                ]
            }
        },
        "/auth": {
            "post": {
                "description": "It takes a token to authenticate yorself to the application",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get a Auth",
                "parameters": [
                    {
                        "description": "Body",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.Login"
                        }
                    }
                ]
            }
        },
        "/deposite": {
            "post": {
                "description": "Make a deposite from an authentic account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Make a deposite",
                "parameters": [
                    {
                        "description": "Body",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/autenticationoperations.Deposite"
                        }
                    }
                ]
            }
        },
        "/transfer": {
            "post": {
                "description": "Transfer between accounts. The account that will make the transfer must be authenticated with a token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Make a transfer",
                "parameters": [
                    {
                        "description": "Body",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/transfer.TransferInput"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Bearer Authorization Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ]
            }
        },
        "/withdraw": {
            "post": {
                "description": "Make a Withdraw from an authentic account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Make a Withdraw",
                "parameters": [
                    {
                        "description": "Body",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/autenticationoperations.Withdraw"
                        }
                    }
                ]
            }
        }
    },
    "definitions": {
        "autenticationoperations.Deposite": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                }
            }
        },
        "autenticationoperations.Withdraw": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                }
            }
        },
        "auth.Login": {
            "type": "object",
            "properties": {
                "cpf": {
                    "type": "string"
                },
                "secret": {
                    "type": "string"
                }
            }
        },
        "entities.Account": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "number"
                },
                "cpf": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "nome": {
                    "type": "string"
                },
                "secret": {
                    "type": "string"
                }
            }
        },
        "entities.AccountInput": {
            "type": "object",
            "properties": {
                "cpf": {
                    "type": "string"
                },
                "nome": {
                    "type": "string"
                },
                "secret": {
                    "type": "string"
                }
            }
        },
        "transfer.TransferInput": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "cpf_destino": {
                    "type": "string"
                }
            }
        }
    }
}